feat: bojo-int Entity 전환 Phase 5 완료 — Step 4개 IF 읽기 JPA 전환
- gen_entities.py: IDENTITY 감지 버그 수정 (@Id 누락 + 타입 String→Long)
- 엔티티 55개 재생성 (PK constraint 파서 수정)
- Factory 3개: DynamicEntityManagerService 주입
- Step 4개 IF 읽기 JPA native query 전환:
  - UseLoadStep: entity getter 직접 사용
  - JejuJewonLoadStep: entityToMap 어댑터 (MERGE 메서드 호환)
  - JejuObsvdataLoadStep: 동일 어댑터 패턴
  - InternalBojoLoadStep: 동일 어댑터 패턴
- ddl-auto: validate 기동 확인 완료

+++ sync-agent-bojo-int/src/main/java/com/sync/agent/bojoint/loader/step/UseLoadStep.java
@@ -220,7 +220,7 @@ public class UseLoadStep implements StepExecutor {
         LegacyResult result = new LegacyResult();
 
         for (IfRsvUseLegacyData row : rows) {
-            Integer ifId = row.getId();
+            Long ifId = row.getId();
             String telno = row.getTelno();
 
             try {
@@ -327,7 +327,7 @@ public class UseLoadStep implements StepExecutor {
         StatusResult result = new StatusResult();
 
         for (IfRsvUseStatusData row : rows) {
-            Integer ifId = row.getId();
+            Long ifId = row.getId();
             Long sn = row.getSn();
 
             try {
